+++ vscode/webviews/promptEditor/fixtures.ts
@@ -1,5 +1,5 @@
+import type { SerializedPromptEditorState } from '@sourcegraph/cody-shared'
 import type { SerializedLexicalNode } from 'lexical'
-import type { SerializedPromptEditorState } from './PromptEditor'
 
 export const FILE_MENTION_EDITOR_STATE_FIXTURE: SerializedPromptEditorState = {
     v: 'lexical-v0',